+++ CHANGELOG.md
@@ -8,3 +8,5 @@
     * Klock `1.7.3` -> `1.8.6`
 
 ## 0.2.1
+
+* Added support of flows: now any `KronScheduler` can be convert to `Flow<DateTime>` using `asFlow` extension